Extract party command
