+++ packages/cli-data/src/modules/party/commands/list.ts
@@ -6,7 +6,7 @@ import { Arguments, CommandModule } from 'yargs';
 
 import { asyncHandler, print } from '@dxos/cli-core';
 
-import { StateManager } from '../../state-manager';
+import { StateManager } from '../../../state-manager';
 import { PartyOptions } from '../party';
 
 export const listCommand = (stateManager: StateManager): CommandModule => ({
