+++ packages/cli-data/src/modules/party/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2020 DXOS.org
+//
+
+export * from './party';
