+++ packages/cli-data/src/modules/party/party.test.ts
@@ -7,7 +7,7 @@ import { createKeyPair, PublicKey } from '@dxos/crypto';
 import { Awaited } from '@dxos/echo-db';
 import { createTestBroker } from '@dxos/signal';
 
-import { StateManager } from '../state-manager';
+import { StateManager } from '../../state-manager';
 import { createCommand } from './commands/create';
 import { inviteCommand } from './commands/invite';
 import { joinCommand } from './commands/join';
